Hemos vuelto a Gestión de productos para añadirle descuentos. Se ha establecido la relacion entre producto - descuento y tipo de descuento. La base de datos se actualiza perfectamente.
También se ha creado los /templates/admin descuentos-productos.html y descuentos-productos-actualizados.html

En /fragments:
se ha añadido tablaAdminProductosDescuentos.html y tablaAdminProductosDescuentosActualizados.html.

Finalmente en /templates: adminProductosDescuentos.html.

Falta añadirle la realción con precio a descuento.

+++ src/main/java/com/eoi/Fruggy/repositorios/RepoDescuento.java
@@ -4,6 +4,8 @@ import com.eoi.Fruggy.entidades.Descuento;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
-public interface RepoDescuento extends JpaRepository<Descuento, Integer> {
+public interface RepoDescuento extends JpaRepository<Descuento, Long> {
 }
